tools/fastbase64: fix size_t -> int conversion warning
write_with_wrapping accumulated a size_t (to_write) into an int
(current_col), which -Wconversion flags as a possibly value-changing
narrowing. The value is in fact bounded: to_write <= remaining ==
wrap_cols - current_col, so current_col + to_write <= wrap_cols and
fits in int. Make the conversion explicit.

+++ tools/fastbase64.cpp
@@ -313,7 +313,7 @@ void CommandLine::write_with_wrapping(std::FILE *fp, const char *data,
       throw std::runtime_error("Failed to write: " +
                                std::string(strerror(errno)));
     }
-    current_col += to_write;
+    current_col += static_cast<int>(to_write);
     i += to_write;
   }
 }
